+++ tools/mcp/dotnet/AzureSDKMCPHub/Azure.SDK.Tools.MCP.Hub/Tools/HelloWorldTool/HelloWorld.cs
@@ -3,7 +3,7 @@ using System.Diagnostics.CodeAnalysis;
 using Azure.SDK.Tools.MCP.Contract;
 using ModelContextProtocol.Server;
 
-namespace Azure.SDK.Tools.MCP.Hub.Tools
+namespace Azure.SDK.Tools.MCP.Hub.Tools.HelloWorldTool
 {
 
     [McpServerToolType, Description("Echoes the message back to the client.")]
