+++ tools/mcp/dotnet/AzureSDKMCPHub/Azure.SDK.Tools.MCP.Hub/Tools/README.md
@@ -8,7 +8,7 @@
   - For now you need to add the attribute yourself.
 - Mark each function with `[McpServerTool]`.
 - Leverage `HelloWorld.cs` for a functioning basic example (more advanced samples to follow)
-- If you require any helper functions or other shared code. Place your code under a _folder_ named the same as your Tool class name.
+- Place your code under a _folder_ named the same as your Tool class name.
   - See `ReleasePlanTool` for an example of this folder structure.  
 
 ## To quickly dev loop on just your class.
